Refactor TdiSdeWrapper class
- Create DPDK, ES2K, and Tofino subclasses of TdiSdeWrapper.

- Move the Tofino-specific packet i/o methods to TofinoSdeWrapper
  and make them unconditional. This eliminates the remaining
  dependencies on the TOFINO_TARGET #define.
