+++ stratum/hal/bin/tdi/dpdk/BUILD
@@ -39,8 +39,8 @@ stratum_cc_binary(
         "//stratum/hal/lib/tdi/dpdk:dpdk_chassis_manager",
         "//stratum/hal/lib/tdi/dpdk:dpdk_hal",
         "//stratum/hal/lib/tdi/dpdk:dpdk_port_manager",
-        "//stratum/hal/lib/tdi/dpdk:dpdk_sde_target",
         "//stratum/hal/lib/tdi/dpdk:dpdk_sde_utils",
+        "//stratum/hal/lib/tdi/dpdk:dpdk_sde_wrapper",
         "//stratum/hal/lib/tdi/dpdk:dpdk_switch",
         "//stratum/hal/lib/tdi:tdi_action_profile_manager",
         "//stratum/hal/lib/tdi:tdi_counter_manager",
