+++ stratum/hal/bin/tdi/es2k/BUILD
@@ -40,8 +40,8 @@ stratum_cc_binary(
         "//stratum/hal/lib/tdi/es2k:es2k_hal",
         "//stratum/hal/lib/tdi/es2k:es2k_node",
         "//stratum/hal/lib/tdi/es2k:es2k_port_manager",
-        "//stratum/hal/lib/tdi/es2k:es2k_sde_target",
         "//stratum/hal/lib/tdi/es2k:es2k_sde_utils",
+        "//stratum/hal/lib/tdi/es2k:es2k_sde_wrapper",
         "//stratum/hal/lib/tdi/es2k:es2k_switch",
         "//stratum/hal/lib/tdi:tdi_action_profile_manager",
         "//stratum/hal/lib/tdi:tdi_counter_manager",
