+++ stratum/hal/bin/tdi/tofino/BUILD
@@ -35,13 +35,12 @@ stratum_cc_binary(
         "//stratum/hal/bin/barefoot:tofino_skip_p4.conf",
         "//stratum/hal/bin/barefoot:tofino_skip_p4_no_bsp.conf",
     ],
-    defines = ["TOFINO_TARGET"],
     deps = [
         "//stratum/hal/bin/tdi:main",
         "//stratum/hal/lib/tdi/tofino:tofino_chassis_manager",
         "//stratum/hal/lib/tdi/tofino:tofino_hal",
         "//stratum/hal/lib/tdi/tofino:tofino_port_manager",
-        "//stratum/hal/lib/tdi/tofino:tofino_sde_target",
+        "//stratum/hal/lib/tdi/tofino:tofino_sde_wrapper",
         "//stratum/hal/lib/tdi/tofino:tofino_sde_utils",
         "//stratum/hal/lib/tdi/tofino:tofino_switch",
         "//stratum/hal/lib/tdi:tdi_action_profile_manager",
