+++ stratum/hal/lib/tdi/es2k/CMakeLists.txt
@@ -18,8 +18,8 @@ target_sources(stratum_tdi_target_o PRIVATE
     es2k_port_manager.cc
     es2k_port_manager.h
     es2k_sde_table_data.cc
-    es2k_sde_target.cc
     es2k_sde_utils.cc
+    es2k_sde_wrapper.cc
     es2k_switch.cc
     es2k_switch.h
 )
